+++ src/game/helpers/animations-creator.js
