+++ src/game/helpers/input-text-handler.js
