+++ src/game/helpers/output-formatter.js
