+++ src/game/helpers/shots-handler.js
