+++ src/game/settings/Preferences.js
