+++ src/game/settings/game-settings.js
